+++ src/common/index.ts
@@ -2,4 +2,4 @@ export * from './hooks';
 export * from './components';
 export * from './Common.constants';
 export * from './Common.utils';
-export * from './Common.types';
+export * from './Common.form';
