+++ src/queries/models/channels/getChannelDetailsModelCreator.ts
@@ -1,4 +1,4 @@
-import { ChannelDetails } from 'common';
+import { ChannelDetails } from '../../types';
 
 //any remodelling needed in the response can be done here
 export const getChannelDetailsModelCreator = (response: ChannelDetails): ChannelDetails => response;
