+++ src/queries/types/channels.ts
@@ -42,3 +42,47 @@ export type SendNotificationParams = {
   config: {};
   request?: any;
 };
+
+export type Alias = {
+  id: number;
+  channel: string;
+  processed: number;
+  alias_address: string;
+  is_alias_verified: number;
+  alias_verification_event: string | null;
+  timestamp: string;
+  initiate_verification_proof: string;
+  verify_verification_proof: string;
+  alias_blockchain_id: string;
+};
+export type ChannelDetails = {
+  activation_status: number;
+  alias_address: string | null;
+  alias_blockchain_id: string | null;
+  alias_verification_event: string | null;
+  attempts: number;
+  blocked: number;
+  channel: string;
+  channel_settings: string;
+  counter: string | null;
+  icon: string;
+  iconV2: string;
+  icon_migration: number;
+  id: number;
+  info: string;
+  initiate_verification_proof: string | null;
+  ipfshash: string;
+  is_alias_verified: number;
+  minimal_channel_settings: string;
+  name: string;
+  payload_migration: number;
+  processed: number;
+  subgraph_attempts: number;
+  subgraph_details: string | null;
+  subscriber_count: number;
+  timestamp: string; // ISO date string
+  url: string;
+  verified_status: number;
+  verify_verification_proof: string | null;
+  aliases: Array<Alias>;
+};
